pyproject.toml
[tool.poetry]
name = "icebreaker"
version = "0.1.0"
description = ""
authors = ["chrisongg"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.14"
langchain-openai = "^0.1.22"
langchain-community = "^0.2.12"
langchainhub = "^0.1.21"
langchain-ollama = "^0.1.1"
python-dotenv = "^1.0.1"
tweepy = "^4.14.0"
flask = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
